Put the spinlock back
SDL_LockAudioDevice will completely stop the audio callbacks
from happening. memcpy of all the samples collected takes too long
and we drop the new sample data. With our own spinlock the callback
is still being called, it just waits on the spinlock. Meaning, there
will be a lot of callbacks called waiting for the spinlock to unlock
but at the end of the day it all works and no samples are lost