Audio recording (using SDL2) (#19)
* init

* update libjepg

* Working implementation. Added my fork of sdl2 as submodule

* fixed bug in stereo

* fdk-aac as submodule

* ready for pr

* Update CMakeLists.txt

Co-authored-by: Joshua 'Pip' Minter <josh@minteronline.com>

* post review changes part1

* add myself to CREDITS.md

* fix oopsie

* audio and encoder separate

* remove unused frameBuffer

* change 1 to true

* clang-format

* grab monitor automaticaly. remove uneeded spinlock

* fix format. update sdl2 submodule

* Put the spinlock back

SDL_LockAudioDevice will completely stop the audio callbacks
from happening. memcpy of all the samples collected takes too long
and we drop the new sample data. With our own spinlock the callback
is still being called, it just waits on the spinlock. Meaning, there
will be a lot of callbacks called waiting for the spinlock to unlock
but at the end of the day it all works and no samples are lost

* Handle removing and adding audio devices

When currently used audio device is running
RS will try opening up another default audio device.
Same happens when a new device is connected.
Reconnection will cause the sample buffer to reset
Causing the first 'config.duration' seconds of the
capture to be out of sync or have bad audio. After
that period everything should be normal.

* post review changes

* fix warrnings

* remove empty line

* Update src/audio_input/audio.c

Co-authored-by: Joshua 'Pip' Minter <josh@minteronline.com>

* Update src/audio_input/audio.c

Co-authored-by: Joshua 'Pip' Minter <josh@minteronline.com>

* Update src/audio_input/audio.c

Co-authored-by: Joshua 'Pip' Minter <josh@minteronline.com>

* Update src/audio_input/audio.c

Co-authored-by: Joshua 'Pip' Minter <josh@minteronline.com>

* Update src/main.c

Co-authored-by: Joshua 'Pip' Minter <josh@minteronline.com>

* Update src/audio_input/audio.c

Co-authored-by: Joshua 'Pip' Minter <josh@minteronline.com>

* Update src/output.c

Co-authored-by: Joshua 'Pip' Minter <josh@minteronline.com>

* Update src/util/circle_static.c

Co-authored-by: Joshua 'Pip' Minter <josh@minteronline.com>

* Update src/util/circle_static.c

Co-authored-by: Joshua 'Pip' Minter <josh@minteronline.com>

* Update src/audio_input/audio.c

Co-authored-by: Joshua 'Pip' Minter <josh@minteronline.com>

* Update src/util/circle_static.c

Co-authored-by: Joshua 'Pip' Minter <josh@minteronline.com>

* Update src/audio_input/audio.c

Co-authored-by: Joshua 'Pip' Minter <josh@minteronline.com>

* post review fix

* Change meaning of audioDeviceName config

auto means select the monitor of the default
output device. system means let SDL2 select
the device (null device name). Any other name
is considered a full device name.
Also another config option was added.
audioSystemFallback which can be 0 or 1.
If set to 1 and opening an audio device fails
a default (null device name) opening will be attempted.

* Proper handling of disconnected device

If connection to the audio device failed then
the rest of the RSAudio setup is not done.
It will be finished if reconnection ever happens.
Also the encoder will not try to read
the non existing sample buffer

* Proper detection of default device

Calling SDL_GetAudioDeviceName with index -1
will return the name of the default output device.

* config option bool. minor tweaks

Co-authored-by: Joshua 'Pip' Minter <josh@minteronline.com>